Moved the page methods to a new class
