+++ modules/checklist/steps/add-logo.php
@@ -2,7 +2,7 @@
 
 namespace Elementor\Modules\Checklist\Steps;
 
-use Elementor\Core\Base\Document;
+use Elementor\Core\DocumentTypes\Page;
 use Elementor\Plugin;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -35,15 +35,15 @@ class Add_Logo extends Step_Base {
 	}
 
 	public function get_cta_url() : string {
-		$link = $this->get_site_settings_url_config();
+		$link = Page::get_site_settings_url_config( true );
 
 		if ( ! $link ) {
-			return $this->get_elementor_create_new_page_url();
+			return Page::get_elementor_create_new_page_url( self::SITE_IDENTITY_TAB );
 		}
 
-		$parsed_url = parse_url( $link );
+		$parsed_url = wp_parse_url( $link );
 		$query_params = [];
-		parse_str( $parsed_url['query'] ?? '', $query_params );
+		wp_parse_str( $parsed_url['query'] ?? '', $query_params );
 		$additional_params = [
 			'active-document' => Plugin::$instance->kits_manager->get_active_id(),
 			'active-tab' => self::SITE_IDENTITY_TAB,
@@ -55,7 +55,7 @@ class Add_Logo extends Step_Base {
 			return $this->wordpress_adapter->add_query_arg( $additional_params, $link );
 		}
 
-		return $this->get_elementor_create_new_page_url();
+		return Page::get_elementor_create_new_page_url();
 	}
 
 	public function get_is_completion_immutable() : bool {
@@ -70,57 +70,9 @@ class Add_Logo extends Step_Base {
 		return 'http://go.elementor.com/app-website-checklist-logo-article';
 	}
 
-	private function get_elementor_create_new_page_url(): string {
-		return $this->wordpress_adapter->add_query_arg( [
-			'active-document' => Plugin::$instance->kits_manager->get_active_id(),
-			'active-tab' => self::SITE_IDENTITY_TAB,
-		], Plugin::$instance->documents->get_create_new_post_url( 'page' ) );
-	}
+	public function page_exists( $params ) : bool {
+		$query = new \WP_Query( $params );
 
-	private function get_site_settings_url_config(): string {
-		$existing_elementor_page = $this->get_elementor_page();
-		return ! empty( $existing_elementor_page )
-			? $this->get_elementor_edit_url( $existing_elementor_page->ID )
-			: $this->get_elementor_create_new_page_url();
-	}
-
-	private function get_elementor_edit_url( int $post_id ): string {
-		$active_kit_id = Plugin::$instance->kits_manager->get_active_id();
-		$document = Plugin::$instance->documents->get( $post_id );
-
-		if ( ! $document ) {
-			return '';
-		}
-
-		return $this->wordpress_adapter->add_query_arg( [ 'active-document' => $active_kit_id ], $document->get_edit_url() );
-	}
-
-	private function get_elementor_page() {
-		$args = [
-			'meta_key' => Document::BUILT_WITH_ELEMENTOR_META_KEY,
-			'sort_order' => 'asc',
-			'sort_column' => 'post_date',
-		];
-		$pages = $this->wordpress_adapter->get_pages( $args );
-
-		if ( empty( $pages ) ) {
-			return null;
-		}
-
-		$show_page_on_front = 'page' === $this->wordpress_adapter->get_option( 'show_on_front' );
-
-		if ( ! $show_page_on_front ) {
-			return $pages[0];
-		}
-
-		$home_page_id = $this->wordpress_adapter->get_option( 'page_on_front' );
-
-		foreach ( $pages as $page ) {
-			if ( (string) $page->ID === $home_page_id ) {
-				return $page;
-			}
-		}
-
-		return $pages[0];
+		return $query->found_posts;
 	}
 }
